[KB-1] GetPod and CurrentPod Init
